extract license class into separate file
